% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdverseEvents.R
\docType{data}
\name{SEVERITY_PROBABILITIES}
\alias{SEVERITY_PROBABILITIES}
\title{Pre defined severity probabilities}
\format{
An object of class \code{list} of length 4.
}
\usage{
SEVERITY_PROBABILITIES
}
\description{
Pre defined severity probabilities
}
\keyword{datasets}
